// types.ts
export interface NavItem {
  label: string;
  href: string;
}

export interface FeatureCard {
  title: string;
  description: string;
  icon: string;
}

export interface SectorCard {
  title: string;
  description: string;
  image: string;
  details: string;
}

export interface NewsletterArchive {
  vol: number;
  date: string;
  title: string;
}